simulation_tacto/requirements.txt
numpy >= 1.18.5
pybullet >= 2.8.1
matplotlib >= 3.1.3
opencv-python >= 3.4.2.17
omegaconf >= 2.0.6
pyrender >= 0.1.43
hydra-core >= 1.0.6
urdfpy >= 0.0.22
scipy
pybulletX >= 0.4.1

# Need a specific version of PyOpenGL (forked & modified by mmatl) for pyrender
# to run headlessly using OSMesa. Otherwise, we will get python ImportError:
# `cannot import name 'OSMesaCreateContextAttribs'`.
pyopengl @ git+https://github.com/mmatl/pyopengl.git@76d1261adee2d3fd99b418e75b0416bb7d2865e6

